@@ -3,76 +3,58 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@aspnet/signalr": "^1.1.4",
-    "@date-io/moment": "2.10.6",
-    "@fortawesome/fontawesome-svg-core": "^1.2.32",
-    "@fortawesome/free-solid-svg-icons": "^5.15.1",
-    "@fortawesome/react-fontawesome": "^0.1.14",
-    "@material-ui/core": "^4.11.2",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.57",
-    "@material-ui/pickers": "3.2.10",
-    "@reduxjs/toolkit": "^1.7.1",
-    "@testing-library/jest-dom": "^5.11.8",
-    "@testing-library/react": "^11.2.2",
-    "@testing-library/user-event": "^12.6.0",
-    "@types/react-redux": "^7.1.15",
-    "@types/webpack-env": "^1.16.0",
-    "axios": "^0.21.4",
-    "classnames": "^2.2.6",
-    "devextreme": "20.2.4",
-    "devextreme-aspnet-data-nojquery": "^2.8.0",
-    "devextreme-react": "20.2.4",
-    "eslint-plugin-sonarjs": "^0.5.0",
+    "@emotion/react": "^11.7.1",
+    "@emotion/styled": "^11.6.0",
+    "@microsoft/signalr": "^6.0.2",
+    "@mui/icons-material": "^5.4.1",
+    "@mui/material": "^5.4.1",
+    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.4.0",
+    "@types/node": "^16.11.24",
+    "@types/react": "^17.0.39",
+    "@types/react-dom": "^17.0.11",
+    "@types/recharts": "^1.8.23",
+    "@types/uuid": "^8.3.4",
+    "@types/webpack-env": "^1.16.3",
+    "axios": "^0.25.0",
+    "date-fns": "^2.28.0",
+    "date-fns-tz": "^1.2.2",
+    "devextreme": "21.2.6",
+    "devextreme-react": "21.2.5",
     "file-saver": "^2.0.5",
-    "i18next": "^19.8.4",
-    "i18next-browser-languagedetector": "^6.0.1",
+    "i18next": "^21.6.11",
+    "i18next-browser-languagedetector": "^6.1.3",
     "i18next-http-backend": "^1.3.2",
     "jwt-decode": "^3.1.2",
-    "material-ui-nested-menu-item": "^1.0.2",
-    "moment": "^2.29.1",
-    "node-sass": "^4.0.0",
-    "notistack": "^1.0.3",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.1",
-    "react-custom-scrollbars": "^4.2.1",
-    "react-dom": "^17.0.1",
-    "react-hook-form": "^7.7.1",
-    "react-i18next": "^11.8.5",
-    "react-moment": "^1.1.1",
-    "react-ratings-declarative": "^3.4.1",
-    "react-redux": "^7.2.2",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.1",
-    "react-spring": "^9.2.1",
-    "react-uuid": "^1.0.2",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
-    "rxjs": "^6.6.3"
+    "notistack": "^2.0.3",
+    "react": "^17.0.2",
+    "react-csv": "^2.2.2",
+    "react-custom-scrollbars-2": "^4.4.0",
+    "react-dom": "^17.0.2",
+    "react-hook-form": "^7.26.1",
+    "react-i18next": "^11.15.4",
+    "react-papaparse": "^4.0.2",
+    "react-router": "^6.2.1",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "5.0.0",
+    "react-spring": "^9.4.2",
+    "rxjs": "^7.5.4",
+    "typescript": "^4.5.5",
+    "uuid": "^8.3.2",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "react-scripts --max_old_space_size=4096 build",
+    "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "lint": "eslint --ext .tsx src --color",
-    "format": "prettier --write src/**/*.{ts,tsx,scss,css,json}",
-    "isready": "npm run format && npm run lint && npm run build",
-    "storybook": "start-storybook -p 6006 -s public",
-    "build-storybook": "build-storybook -s public"
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app",
-    "overrides": [
-      {
-        "files": [
-          "**/*.stories.*"
-        ],
-        "rules": {
-          "import/no-anonymous-default-export": "off"
-        }
-      }
+    "extends": [
+      "react-app",
+      "react-app/jest"
     ]
   },
   "browserslist": {
@@ -88,39 +70,12 @@
     ]
   },
   "devDependencies": {
-    "@react-theming/storybook-addon": "^1.1.1",
-    "@storybook/addon-actions": "^6.2.9",
-    "@storybook/addon-essentials": "^6.4.14",
-    "@storybook/addon-links": "^6.2.9",
-    "@storybook/node-logger": "^6.2.9",
-    "@storybook/preset-create-react-app": "^3.1.7",
-    "@storybook/react": "^6.4.14",
-    "@types/jest": "^26.0.19",
-    "@types/node": "^14.14.19",
-    "@types/query-string": "^6.3.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-router-dom": "^5.1.7",
-    "@typescript-eslint/eslint-plugin": "4.11.1",
-    "@typescript-eslint/parser": "4.11.1",
-    "babel-eslint": "10.1.0",
-    "eslint": "^7.17.0",
-    "eslint-config-prettier": "7.1.0",
-    "eslint-config-react-app": "6.0.0",
-    "eslint-import-resolver-typescript": "2.3.0",
-    "eslint-loader": "4.0.2",
-    "eslint-plugin-flowtype": "^5.2.0",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-prettier": "3.3.0",
-    "eslint-plugin-react": "7.22.0",
-    "eslint-plugin-react-hooks": "4.2.0",
-    "eslint-plugin-you-dont-need-momentjs": "^1.6.0",
-    "prettier": "2.2.1",
-    "prettier-eslint": "12.0.0",
-    "prettier-eslint-cli": "5.0.0",
-    "storybook-addon-material-ui": "^0.9.0-alpha.24",
-    "typescript": "^4.1.3",
-    "webpack-bundle-analyzer": "^4.3.0"
+    "@types/file-saver": "^2.0.5",
+    "@typescript-eslint/eslint-plugin": "^5.11.0",
+    "@typescript-eslint/parser": "^5.11.0",
+    "eslint": "^8.8.0",
+    "eslint-config-google": "^0.14.0",
+    "eslint-plugin-react": "^7.28.0",
+    "eslist": "^1.0.0-beta.1"
   }
 }
