@@ -1,63 +1,55 @@
-import React, { useState } from 'react';
-import Grid from '@material-ui/core/Grid';
-import Card from '@material-ui/core/Card';
-import CardContent from '@material-ui/core/CardContent';
-import { makeStyles } from '@material-ui/core/styles';
-import { useTransition, config } from 'react-spring';
+import {Grid, Card, CardContent} from '@mui/material';
+import React, {useState, useEffect, useContext} from 'react';
+import {Navigate} from 'react-router';
 
-import { LoginForm, RecoverPasswordForm } from '.';
+import {LoginForm, RecoverPasswordForm} from '.';
 import SmallLogo from '../../shared/assets/images/logo-small.png';
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    width: '100%',
-    height: '100%',
-  },
-  card: {
-    width: 325,
-    height: 350,
-    padding: 15,
-  },
-  card__content: {
-    height: '100%',
-  },
-  logo: {
-    marginLeft: 10,
-  }
-}));
+import {UserContext} from '../../shared/context/user';
 
 export const Login = () => {
-  const classes = useStyles();
   const [inRecoveryMode, setInRecoveryMode] = useState(false);
+  const [isLoggedIn, setIsLoggedIn] = useState(false);
+  const userContext = useContext(UserContext);
 
-  const transitions = useTransition(inRecoveryMode, {
-    from: { position: 'absolute', opacity: 0 },
-    enter: { opacity: 1 },
-    leave: { opacity: 0 },
-    config: config.default,
-  });
+  useEffect(() => {
+    (async () => {
+      const result = await userContext.isUserLoggedIn();
+      setIsLoggedIn(result);
+    })();
+  }, []);
+
+  if (isLoggedIn) {
+    return <Navigate to="/" replace />;
+  };
 
   return (
     <Grid
       container
-      justify="center"
+      justifyContent="center"
       alignItems="center"
-      className={classes.root}>
-      <Card className={classes.card}>
-        <CardContent className={classes.card__content}>
-          <Grid container direction="column" justify="space-between">
-            <Grid item>
-              <img src={SmallLogo} width={250} alt="EmbyStat logo" className={classes.logo} />
+      sx={{
+        width: '100vw',
+        height: '100vh',
+      }}>
+      <Card>
+        <CardContent
+          sx={{
+            paddingTop: '20px',
+          }}>
+          <Grid
+            container
+            direction="column"
+            justifyContent="space-between"
+          >
+            <Grid item container justifyContent="center">
+              <img src={SmallLogo} width={250} alt="EmbyStat logo" style={{marginBottom: 10}}/>
             </Grid>
-            <Grid item className={classes.card__content}>
-              {transitions((style, item) =>
-                item
-                  ? <RecoverPasswordForm style={{ ...style, height: '235px', width: '263px' }} openLoginForm={() => setInRecoveryMode((prev) => !prev)} />
-                  : <LoginForm style={{ ...style, height: '235px', width: '263px' }} openForgotPasswordForm={() => setInRecoveryMode((prev) => !prev)} />
-              )}
+            <Grid item>
+              {
+                inRecoveryMode ?
+                  <RecoverPasswordForm openLoginForm={() => setInRecoveryMode((prev) => !prev)} /> :
+                  <LoginForm openForgotPasswordForm={() => setInRecoveryMode((prev) => !prev)} />
+              }
             </Grid>
           </Grid>
         </CardContent>
