@@ -1,7 +1,7 @@
 import i18n from 'i18next';
-import { initReactI18next } from 'react-i18next';
-import HttpApi from 'i18next-http-backend';
 import LanguageDetector from 'i18next-browser-languagedetector';
+import HttpApi from 'i18next-http-backend';
+import {initReactI18next} from 'react-i18next';
 
 let translationFilePath = '/locales/{{lng}}.json';
 if (process.env.NODE_ENV === 'development' && module.hot) {
@@ -23,37 +23,36 @@ i18n
         'path',
         'subdomain',
       ],
-      checkWhitelist: true,
     },
     backend: {
       loadPath: translationFilePath,
     },
-    whitelist: [
-      'cs-CZ',
-      'da-DK',
-      'de-DE',
-      'el-GR',
-      'en-US',
-      'es-ES',
-      'fi-FI',
-      'fr-FR',
-      'hu-HU',
-      'it-IT',
-      'nl-NL',
-      'no-NO',
-      'pl-PL',
-      'pt-BR',
-      'pt-PT',
-      'ro-RO',
-      'sv-SE',
+    supportedLngs: [
+      'cs',
+      'da',
+      'de',
+      'el',
+      'en',
+      'es',
+      'fi',
+      'fr',
+      'hu',
+      'it',
+      'nl',
+      'no',
+      'pl',
+      'pt',
+      'pt',
+      'ro',
+      'sv',
     ],
-    fallbackLng: 'en-US',
+    fallbackLng: 'en',
     debug: false,
     interpolation: {
       escapeValue: false,
     },
     react: {
-      wait: true,
+      useSuspense: true,
     },
   });
 
