@@ -1,55 +1,34 @@
-import './i18n';
-import 'moment/locale/cs';
-import 'moment/locale/da';
-import 'moment/locale/de';
-import 'moment/locale/el';
-import 'moment/locale/es';
-import 'moment/locale/fi';
-import 'moment/locale/fr';
-import 'moment/locale/hu';
-import 'moment/locale/it';
-import 'moment/locale/nl';
-import 'moment/locale/nb';
-import 'moment/locale/pl';
-import 'moment/locale/pt-br';
-import 'moment/locale/pt';
-import 'moment/locale/ro';
-import 'moment/locale/sv';
-
-import './index.scss';
-
-import React, { Suspense } from 'react';
+import React, {Suspense} from 'react';
 import ReactDOM from 'react-dom';
-import { Provider } from 'react-redux';
-import { BrowserRouter } from 'react-router-dom';
+import {BrowserRouter} from 'react-router-dom';
 
+import './index.css';
+import './i18n';
 import App from './App';
-import * as serviceWorker from './serviceWorker';
-import store from './store';
-import PageLoader from './shared/components/pageLoader';
-import { SettingsContextProvider } from './shared/context/settings';
+import reportWebVitals from './reportWebVitals';
+import {EsPageLoader} from './shared/components/esPageLoader';
+import {SettingsContextProvider} from './shared/context/settings';
+import {UserContextProvider} from './shared/context/user';
+import {JobsContextProvider} from './shared/context/jobs';
 
-function render() {
-  ReactDOM.render(
-    // <React.StrictMode>
-    <Provider store={store}>
-      <SettingsContextProvider>
-        <Suspense fallback={PageLoader}>
-          <BrowserRouter>
-            <App />
-          </BrowserRouter>
-        </Suspense>
-      </SettingsContextProvider>
-    </Provider>,
-    // </React.StrictMode>,
-    document.getElementById('root')
-  );
-}
+ReactDOM.render(
+  <React.StrictMode>
+    <SettingsContextProvider>
+      <JobsContextProvider>
+        <UserContextProvider>
+          <Suspense fallback={<EsPageLoader />}>
+            <BrowserRouter>
+              <App />
+            </BrowserRouter>
+          </Suspense>
+        </UserContextProvider>
+      </JobsContextProvider>
+    </SettingsContextProvider>
+  </React.StrictMode>,
+  document.getElementById('root'),
+);
 
-render();
-
-if (process.env.NODE_ENV === 'development' && module.hot) {
-  module.hot.accept('./App', render);
-}
-
-serviceWorker.unregister();
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
