@@ -1,60 +1,29 @@
-import React, { useState } from 'react'
-import { animated } from 'react-spring';
-import Grid from '@material-ui/core/Grid';
-import Button from '@material-ui/core/Button';
-import CircularProgress from '@material-ui/core/CircularProgress';
-import { makeStyles } from '@material-ui/core/styles';
-import { useTranslation } from 'react-i18next';
-import { useForm } from 'react-hook-form';
+import React, {useState} from 'react';
+import {useForm} from 'react-hook-form';
+import {useTranslation} from 'react-i18next';
 
-import { resetPassword } from '../../shared/services/AccountService';
+import {Button, CircularProgress, Grid, Typography} from '@mui/material';
+
+import {EsTextInput} from '../../shared/components/esTextInput';
+import {resetPassword} from '../../shared/services/accountService';
 import SnackbarUtils from '../../shared/utils/SnackbarUtilsConfigurator';
-import { EsTextInput } from '../../shared/components/esTextInput';
-
-const useStyles = makeStyles((theme) => ({
-  card__content: {
-    height: '100%',
-  },
-  button: {
-    marginTop: 15,
-    height: 36,
-    width: '100%'
-  },
-  input__padding: {
-    marginBottom: 30,
-  },
-  input__root: {
-    marginTop: 0,
-  },
-  error_message: {
-    color: theme.palette.error.main,
-    fontSize: '0.8rem',
-    lineHeight: '0.8rem',
-    marginTop: 10,
-    height: 20,
-  },
-  recover__button: {
-    color: "#d3d3d3",
-  }
-}));
 
 interface Props {
-  style: any;
   openLoginForm: () => void;
 }
 
 export const RecoverPasswordForm = (props: Props) => {
-  const { style, openLoginForm } = props;
-  const classes = useStyles();
-  const { t } = useTranslation();
+  const {openLoginForm} = props;
+  const {t} = useTranslation();
   const [isLoading, setIsLoading] = useState(false);
   const [username, setUsername] = useState<string>('');
+  const [resetFailed, setResetFailed] = useState(false);
 
   const usernameChanged = (value: string) => {
     setUsername(value);
   };
 
-  const { register, getValues } = useForm({
+  const {register, getValues} = useForm({
     mode: 'onBlur',
     defaultValues: {
       username: '',
@@ -64,59 +33,76 @@ export const RecoverPasswordForm = (props: Props) => {
   const recoverPassword = () => {
     setIsLoading(true);
     resetPassword(username)
-      .then((result => {
+      .then(((result) => {
+        setResetFailed(!result);
         if (result) {
           SnackbarUtils.success(t('LOGIN.RESETCOMPLETED'));
           openLoginForm();
-        } else {
-          SnackbarUtils.error(t('JOB.RESETFAILED'));
         }
       }))
       .finally(() => {
         setIsLoading(false);
-      })
-  }
+      });
+  };
 
   const usernameRegister = register('username');
 
   return (
-    <animated.div style={{ ...style }}>
-      <Grid container direction="column" justify="flex-end" spacing={1} className={classes.card__content}>
-        <form>
-          <Grid item>
-            <EsTextInput
-              inputRef={usernameRegister}
-              label={t('SETTINGS.ACCOUNT.USERNAME')}
-              onChange={usernameChanged}
-              defaultValue={getValues('username')}
-            />
+    <form>
+      <Grid
+        container
+        direction="column"
+        spacing={4}
+      >
+        <Grid item>
+          <Grid container direction="column">
+            <Grid item>
+              <EsTextInput
+                inputRef={usernameRegister}
+                label={t('SETTINGS.ACCOUNT.USERNAME')}
+                onChange={usernameChanged}
+                defaultValue={getValues('username')}
+              />
+            </Grid>
           </Grid>
-          <Grid item xs={12}>
-            <Button
-              variant="contained"
-              onClick={recoverPassword}
-              className={classes.button}
-              disabled={isLoading || username === ''}
-              color="primary">
-              {
-                isLoading
-                  ? <CircularProgress size={16} className={classes.recover__button} />
-                  : t('LOGIN.RECOVERPASSWORD')
-              }
-            </Button>
+        </Grid>
+
+        <Grid item>
+          <Grid container spacing={1} direction="column" alignItems="center">
+            <Grid item
+              sx={{
+                marginBottom: resetFailed ? 0 : '23px',
+              }}>
+              <Typography variant="body1" color="error">{resetFailed ? t('LOGIN.RESETFAILED') : null}</Typography>
+            </Grid>
+            <Grid item container direction="row" justifyContent="space-between">
+              <Grid item>
+                <Button
+                  variant="text"
+                  onClick={openLoginForm}
+                  disabled={isLoading}
+                  color="secondary"
+                >
+                  {t('COMMON.BACK')}
+                </Button>
+              </Grid>
+              <Grid item>
+                <Button
+                  variant="contained"
+                  onClick={recoverPassword}
+                  disabled={isLoading || username === ''}
+                  color="primary">
+                  {
+                    isLoading ?
+                      <CircularProgress size={16} sx={{color: '#d3d3d3'}} /> :
+                      t('LOGIN.RECOVERPASSWORD')
+                  }
+                </Button>
+              </Grid>
+            </Grid>
           </Grid>
-          <Grid item xs={12}>
-            <Button
-              variant="text"
-              onClick={openLoginForm}
-              className={classes.button}
-              disabled={isLoading}
-              color="secondary">
-              {t('COMMON.BACK')}
-            </Button>
-          </Grid>
-        </form>
+        </Grid>
       </Grid>
-    </animated.div>
-  )
-}
+    </form>
+  );
+};
