@@ -1,65 +1,92 @@
-import Grid from '@material-ui/core/Grid'
-import Typography from '@material-ui/core/Typography'
-import React, { forwardRef, useContext, useEffect, useState } from 'react'
-import { useTranslation } from 'react-i18next'
+import React, {forwardRef, useContext, useEffect, useImperativeHandle} from 'react';
+import {useTranslation} from 'react-i18next';
 
-import { LibraryStepProps, ValidationHandleWithSave } from '.'
-import LibrarySelector from '../../../shared/components/librarySelector'
-import { LibraryContainer } from '../../../shared/models/settings'
-import { WizardContext } from '../Context/WizardState'
+import {Box, Grid, Stack, Typography} from '@mui/material';
 
-export const ConfigureLibrary = forwardRef<ValidationHandleWithSave, LibraryStepProps>((props, ref) => {
-  const { type } = props;
-  const [typeNumber, setTypeNumber] = useState(0);
-  const [selectedLibraries, setSelectedLibraries] = useState<LibraryContainer[]>([]);
-  const { wizard, fullServerUrl, setMovieLibraries, setShowLibraries } = useContext(WizardContext);
-  const { t } = useTranslation();
+import {EsLibraryCard} from '../../../shared/components/esLibrarySelector';
+import {LibrariesContext, LibrariesContextProvider} from '../../../shared/context/library';
+import {WizardContext} from '../../../shared/context/wizard/WizardState';
+import {Library} from '../../../shared/models/library';
+import {fetchLibraries as fetchMovieLibraries} from '../../../shared/services/movieService';
+import {fetchLibraries as fetchShowLibraries} from '../../../shared/services/showService';
+import {LibraryStepProps, ValidationHandle} from '../Interfaces';
 
-  useEffect(() => {
-    if (type === 'movie') {
-      setTypeNumber(1);
-      setSelectedLibraries(wizard.movieLibraries);
-      return;
-    }
-
-    if (type === 'show') {
-      setTypeNumber(2);
-      setSelectedLibraries(wizard.showLibraries);
-    }
-  }, [type, typeNumber, wizard.allLibraries, wizard.movieLibraries, wizard.showLibraries]);
-
-  const saveLibrariesToWizard = (libraries: LibraryContainer[]): void => {
-    if (type === 'movie') {
-      setMovieLibraries(libraries);
-      return;
-    }
-
-    if (type === 'show') {
-      setShowLibraries(libraries);
-      return;
-    }
+export const ConfigureLibrary =
+forwardRef<ValidationHandle, LibraryStepProps>(function ConfigureLibrary(props, ref) {
+  // eslint-disable-next-line react/prop-types
+  const {type} = props;
+  const {setMovieLibraryIds, setShowLibraryIds} = useContext(WizardContext);
+  if (type === 'movie') {
+    return (
+      <LibrariesContextProvider>
+        <ConfigureLibraryContainer
+          fetch={fetchMovieLibraries}
+          type="COMMON.MOVIE"
+          push={setMovieLibraryIds}
+          ref={ref}
+        />
+      </LibrariesContextProvider>
+    );
   }
 
   return (
-    <Grid container direction="column" spacing={2}>
-      <Grid item>
-        <Typography variant="h4" color="primary">
-          {t('WIZARD.CONFIGURELIBRARIES', { type: type })}
-        </Typography>
-      </Grid>
-      <Grid item>
-        <Typography variant="body1">
-          {t('WIZARD.SELECTLIBRARIES', { type: t(`COMMON.${type.toUpperCase()}`) })}
-        </Typography>
-      </Grid>
-      <Grid item>
-        <LibrarySelector
-          allLibraries={wizard.allLibraries}
-          libraries={selectedLibraries}
-          address={fullServerUrl}
-          saveList={(libraries) => saveLibrariesToWizard(libraries)}
-        />
-      </Grid>
-    </Grid>
-  )
-})
+    <LibrariesContextProvider>
+      <ConfigureLibraryContainer
+        fetch={fetchShowLibraries}
+        type="COMMON.SHOW"
+        push={setShowLibraryIds}
+        ref={ref} />
+    </LibrariesContextProvider>
+  );
+});
+
+type Props = {
+  fetch: () => Promise<Library[]>;
+  type: 'COMMON.SHOW' | 'COMMON.MOVIE';
+  push: (libraryIds: string[]) => Promise<void>;
+}
+
+export const ConfigureLibraryContainer =
+forwardRef<ValidationHandle, Props>(function ConfigureLibraryContainer(props, ref) {
+  // eslint-disable-next-line react/prop-types
+  const {fetch, type, push} = props;
+  const {t} = useTranslation();
+  const {load, libraries, save} = useContext(LibrariesContext);
+  const {wizard} = useContext(WizardContext);
+
+  useEffect(() => {
+    (async () => {
+      load(fetch);
+    })();
+  }, []);
+
+  useImperativeHandle(ref, () => ({
+    async validate(): Promise<boolean> {
+      await save(push);
+      return Promise.resolve(true);
+    },
+  }));
+
+  return (
+    <Stack spacing={2}>
+      <Typography variant="h4" color="primary">
+        {t('WIZARD.LIBRARIES.TITLE', {type: t(type)})}
+      </Typography>
+      <Typography variant="body1">
+        {t('SETTINGS.LIBRARIES.CONTENT', {type: t(type)})}
+      </Typography>
+      <Box>
+        <Grid container spacing={1}>
+          {
+            libraries.map((lib) => (
+              <EsLibraryCard
+                key={lib.id}
+                library={lib}
+                fallbackServerUrl={wizard.address}
+              />))
+          }
+        </Grid>
+      </Box>
+    </Stack>
+  );
+});
