@@ -1,6 +1,6 @@
 export interface Card {
   title: string;
   value: string;
-  type: string;
+  type: 'text' | 'time' | 'size';
   icon: string;
 }
