@@ -1,6 +1,6 @@
-export * from './Card';
 export * from './Chart';
-export * from './Media';
+export * from './Card';
 export * from './TopCard';
-export * from './LabelValuePair';
+export * from './Media';
 export * from './TablePage';
+export * from './LabelValuePair';
