@@ -1,26 +1,10 @@
+import {FilterOperation} from './';
+
 export interface ActiveFilter {
   field: string;
   fieldLabel: string;
-  operation:
-  | 'any'
-  | '!any'
-  | 'null'
-  | '!null'
-  | 'between'
-  | 'in'
-  | 'contains'
-  | '!contains'
-  | 'empty'
-  | '=='
-  | '!='
-  | '<'
-  | '>'
-  | 'startsWith'
-  | 'endsWith';
-  operationLabel: string;
+  fieldValue: string;
+  operation: FilterOperation;
   value: string;
-  valueLabel: string;
   id: string;
-  visible: boolean;
-  unit?: string;
 }
