@@ -3,42 +3,38 @@ export interface FilterItem {
   value: string;
 }
 
-export interface FilterType {
+export interface FilterOperation {
   operation:
-    | "any"
-    | "!any"
-    | "null"
-    | "!null"
-    | "between"
-    | "in"
-    | "contains"
-    | "!contains"
-    | "empty"
-    | "=="
-    | "!="
-    | "<"
-    | ">"
-    | "startsWith"
-    | "endsWith";
+    | 'any'
+    | '!any'
+    | 'null'
+    | '!null'
+    | 'between'
+    | 'in'
+    | 'contains'
+    | '!contains'
+    | 'empty'
+    | '=='
+    | '!='
+    | '<'
+    | '>'
+    | 'startsWith'
+    | 'endsWith';
   label: string;
-  type: "txt" | "none" | "dropdown" | "range" | "number" | "date" | "dateRange";
-  itemType?: "url" | "static";
+  type: 'text' | 'none' | 'dropdown' | 'range' | 'number';
+  itemType?: 'url' | 'static';
   items?: FilterItem[];
   itemUrl?: string;
-  min?: number;
-  max?: number;
-  step?: number;
   unit?: string;
   id: string;
   open: boolean;
-  placeholder?: string;
 }
 
 export interface FilterDefinition {
   field: string;
   label: string;
   title: string;
-  types: FilterType[];
+  operations: FilterOperation[];
   open: boolean;
   id: string;
 }
