@@ -1,3 +1,3 @@
-export * from './FilterDefinition';
 export * from './ActiveFilter';
+export * from './FilterDefinition';
 export * from './FilterValues';
