@@ -1,4 +1,3 @@
 export * from './Jobs';
 export * from './JobLogLine';
-export * from './JobsContainer';
-export * from './EnhancedJobLogLine';
+export * from './JobProgress';
