@@ -1,6 +1,3 @@
-export * from './MoviePoster';
 export * from './MovieStatistics';
-export * from './ShortMovie';
-export * from './SuspiciousMovie';
-export * from './MovieStatisticsContainer';
 export * from './MovieRow';
+export * from './Movie';
