@@ -1,14 +1,11 @@
 export interface MediaServerSettings {
-  serverName: string;
+  name: string;
   apiKey: string;
-  serverAddress: string;
-  serverPort: number;
+  address: string;
   authorizationScheme: string;
-  serverProtocol: number;
-  serverType: number;
+  type: number;
   userId: string;
-  serverBaseUrl: string;
-  serverId: string;
+  id: string;
 }
 
 export interface TmdbSettings {
@@ -24,8 +21,6 @@ export interface Settings {
   toShortMovie: number;
   toShortMovieEnabled: boolean;
   keepLogsCount: number;
-  movieLibraries: LibraryContainer[];
-  showLibraries: LibraryContainer[];
   autoUpdate: boolean;
   updateTrain: number;
   updateInProgress: boolean;
@@ -44,47 +39,5 @@ export interface LibraryContainer {
   id: string;
   lastSynced: Date | null;
   name: string;
-}
-
-export const initialTmdbSettings: TmdbSettings = {
-  apiKey: '',
-  lastUpdate: null
-}
-
-export const intialMediaServerSettings: MediaServerSettings = {
-  serverAddress: "",
-  serverName: "",
-  serverPort: 8096,
-  serverProtocol: 0,
-  apiKey: "",
-  serverType: 0,
-  authorizationScheme: '',
-  serverBaseUrl: '',
-  serverId: '',
-  userId: ''
-}
-
-export const initialSettingsState: Settings = {
-  appName: '',
-  autoUpdate: true,
-  configDir: '',
-  dataDir: '',
-  enableRollbarLogging: false,
-  id: '',
-  keepLogsCount: 10,
-  language: 'en',
-  logDir: '',
-  movieLibraries: [],
-  showLibraries: [],
-  noUpdates: false,
-  toShortMovie: 0,
-  toShortMovieEnabled: false,
-  updateInProgress: false,
-  updateTrain: 0,
-  version: '',
-  wizardFinished: false,
-  tmdb: initialTmdbSettings,
-  mediaServer: intialMediaServerSettings,
-  isLoaded: false
 }
 
