@@ -1,10 +1,8 @@
-import { Card, Chart, TopCard } from '../common';
-import { PersonStatistics } from '../person';
+import {Card, Chart, TopCard} from '../common';
 
 export interface ShowStatistics {
   cards: Card[];
   topCards: TopCard[];
   barCharts: Chart[];
   pieCharts: Chart[];
-  people: PersonStatistics;
 }
