@@ -1,5 +1,5 @@
-import { axiosInstance } from './axiosInstance';
-import { Job } from '../models/jobs';
+import {axiosInstance} from './axiosInstance';
+import {Job} from '../models/jobs';
 
 const domain = 'job/';
 
@@ -8,19 +8,13 @@ export const getAllJobs = (): Promise<Job[]> => {
 };
 
 export const fireJob = (id: string): Promise<void> => {
-  return axiosInstance.post(`${domain}fire/${id}`).then((response) => response.data);
+  return axiosInstance.post(`${domain}fire/${id}`);
 };
 
 export const updateTrigger = (id: string, cron: string): Promise<void> => {
-  return axiosInstance.patch(
-    `${domain}${id}`,
-    {},
-    {
-      params: { cron },
-    }
-  );
+  return axiosInstance.patch(`${domain}${id}`, {}, {params: {cron}} );
 };
 
 export const getJobById = (id: string): Promise<Job> => {
   return axiosInstance.get<Job>(`${domain}${id}`).then((response) => response.data);
-}
+};
