@@ -1,7 +1,7 @@
-import { axiosInstance } from './axiosInstance';
-import { saveAs } from 'file-saver';
+import {saveAs} from 'file-saver';
 
-import { LogFile } from '../models/logs';
+import {LogFile} from '../models/logs';
+import {axiosInstance} from './axiosInstance';
 
 const domain = 'log/';
 
@@ -10,20 +10,22 @@ export const getLogList = (): Promise<LogFile[]> => {
 };
 
 export const downloadLogFile = (filename: string, anonymised: boolean): void => {
-  axiosInstance.get(`${domain}download/${filename}`, { params: { anonymous: anonymised } }).then((response) => saveFile(response));
+  axiosInstance
+    .get(`${domain}download/${filename}`, {params: {anonymous: anonymised}})
+    .then((response) => saveFile(response));
 };
 
-const saveFile = (response) => {
+const saveFile = (response: any) => {
   if (response.status !== 200) {
     throw new Error(response);
   }
 
   const filename = response.headers['content-disposition']
     .split(';')
-    .find((n) => n.includes('filename='))
+    .find((n: string) => n.includes('filename='))
     .replace('filename=', '')
     .replace(/"/g, '')
     .trim();
-  var blob = new Blob([response.data], { type: "text/plain;charset=utf-8" });
+  const blob = new Blob([response.data], {type: 'text/plain;charset=utf-8'});
   saveAs(blob, filename);
 };
