@@ -1,10 +1,10 @@
+import {AxiosResponse} from 'axios';
 import i18n from 'i18next';
-import { AxiosResponse } from 'axios';
 
-import { axiosInstance } from './axiosInstance';
-import { Settings } from '../models/settings';
-import { Language } from '../models/language';
+import {Language} from '../models/language';
+import {Settings} from '../models/settings';
 import SnackbarUtils from '../utils/SnackbarUtilsConfigurator';
+import {axiosInstance} from './axiosInstance';
 
 const domain = 'settings/';
 
@@ -14,11 +14,11 @@ export const getSettings = async (): Promise<Settings> => {
 };
 
 export const updateSettings = async (userSettings: Settings): Promise<Settings | void> => {
-  const config = { headers: { 'Content-Type': 'application/json' } };
+  const config = {headers: {'Content-Type': 'application/json'}};
   return await axiosInstance.put<Settings>(`${domain}`, userSettings, config)
     .then((response) => {
       if (response.status === 200) {
-        SnackbarUtils.success(i18n.t('SETTINGS.SAVED'))
+        SnackbarUtils.success(i18n.t('SETTINGS.SAVED'));
       }
       return response.data;
     })
@@ -30,42 +30,3 @@ export const updateSettings = async (userSettings: Settings): Promise<Settings |
 export const getLanguages = async (): Promise<AxiosResponse<Language[]>> => {
   return await axiosInstance.get<Language[]>(`${domain}languages`);
 };
-
-// export const useGetlanguage = () => {
-//   const [languageList, setLanguageList] = useState<Language[]>([]);
-//   const [isLoading, setIsLoading] = useState(false);
-//   const [inError, setInError] = useState(false);
-
-//   useEffect(() => {
-//     let didCancel = false;
-
-//     const fetchData = async () => {
-//       try {
-//         setIsLoading(true);
-//         const result = await axiosInstance.get<Language[]>(`${domain}languages`);
-
-//         if (!didCancel) {
-//           setLanguageList(result.data);
-//         }
-//       } catch (error) {
-//         if (!didCancel) {
-//           setInError(true);
-//         }
-//       }
-//       finally {
-//         setIsLoading(false);
-//       }
-//     };
-
-//     fetchData();
-
-//     return () => {
-//       didCancel = true;
-//     };
-//   }, []);
-
-//   return {
-//     languageList, isLoading, inError
-//   }
-// }
-
